Moves PurchaserInfo
